+++ Core/Interoperability/CoreUrlProcessor.cs
@@ -1,0 +1,12 @@
+using System.Diagnostics;
+
+namespace Unity.Cloud.Common
+{
+    public class CoreUrlProcessor : IUrlProcessor
+    {
+        public void ProcessURL(string url)
+        {
+            Process.Start(url);
+        }
+    }
+}