+++ Core/Networking/Http/IHttpClient.cs
@@ -16,24 +16,18 @@ namespace Unity.Cloud.Common
         public TimeSpan Timeout { get; set; }
 
         /// <summary>
-        /// Send an asynchronous HTTP request.
+        /// Sends an asynchronous HTTP request.
         /// </summary>
         /// <param name="request">The request to be sent.</param>
-        /// <param name="cancellationToken">Optional cancellation token that will try to cancel the operation.</param>
+        /// <param name="progress">The progress provider.</param>
+        /// <param name="completionOption">When the operation should complete.</param>
+        /// <param name="cancellationToken">Cancellation token that will try to cancel the operation.</param>
         /// <returns>A task that will hold the HttpResponseMessage once the request is completed</returns>
         /// <exception cref="ArgumentNullException">Thrown when the request is null.</exception>
         /// <exception cref="HttpRequestException">Thrown when an HTTP response can't be obtained from the server.</exception>
         /// <exception cref="OperationCanceledException">Thrown when the request is cancelled by a cancellation token.</exception>
         /// <exception cref="TimeoutException">Thrown when the request failed due to timeout.</exception>
-        Task<HttpResponseMessage> SendAsync(HttpRequestMessage request, CancellationToken cancellationToken = default);
-
-        /// <summary>
-        /// Send an asynchronous file download request.
-        /// </summary>
-        /// <param name="request">The request to be sent.</param>
-        /// <param name="downloadFilePath">Optional path to save downloaded files.</param>
-        /// <param name="cancellationToken">Optional cancellation token that will try to cancel the operation.</param>
-        /// <returns>A task that will hold the HttpResponseMessage once the request is completed</returns>
-        Task<HttpResponseMessage> DownloadFileAsync(HttpRequestMessage request, string downloadFilePath, CancellationToken cancellationToken = default);
+        Task<HttpResponseMessage> SendAsync(HttpRequestMessage request, HttpCompletionOption completionOption, IProgress<HttpProgress> progress,
+            CancellationToken cancellationToken);
     }
 }