+++ Core/Networking/Http/IServiceHttpClient.cs
@@ -15,6 +15,8 @@ namespace Unity.Cloud.Common
         /// </summary>
         /// <param name="request">The request to be sent.</param>
         /// <param name="options">The options for the client.</param>
+        /// <param name="completionOption">When the operation should complete.</param>
+        /// <param name="progress">The progress provider.</param>
         /// <param name="cancellationToken">Optional cancellation token that will try to cancel the operation.</param>
         /// <returns>A task that will hold the HttpResponseMessage once the request is completed</returns>
         /// <exception cref="ArgumentNullException">Thrown when the request is null.</exception>
@@ -22,6 +24,6 @@ namespace Unity.Cloud.Common
         /// <exception cref="OperationCanceledException">Thrown when the request is cancelled by a cancellation token.</exception>
         /// <exception cref="TimeoutException">Thrown when the request failed due to timeout.</exception>
         public Task<HttpResponseMessage> SendAsync(HttpRequestMessage request, ServiceHttpClientOptions options,
-            CancellationToken cancellationToken = default);
+            HttpCompletionOption completionOption, IProgress<HttpProgress> progress, CancellationToken cancellationToken);
     }
 }