+++ Runtime/Interoperability/UrlProcessor.cs
@@ -1,0 +1,12 @@
+using UnityEngine;
+
+namespace Unity.Cloud.Common.Runtime
+{
+    public class UrlProcessor : IUrlProcessor
+    {
+        public void ProcessURL(string url)
+        {
+            Application.OpenURL(url);
+        }
+    }
+}